+++ internal/controller/statefulset_webhook.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"slices"
 	"strings"
+	"text/template"
 	"time"
 
 	rm "cloud.google.com/go/resourcemanager/apiv3"
@@ -53,13 +54,6 @@ type StatefulsetMutator struct {
 	GroupConfigs []AccessGroupConfig
 }
 
-type AccessGroupConfig struct {
-	Name            string        `yaml:"name"`
-	ProjectTemplate string        `yaml:"projectTemplate"`
-	MaxDuration     time.Duration `yaml:"maxDuration"`
-	ReasonRequired  bool          `yaml:"reasonRequired"`
-}
-
 func (m *StatefulsetMutator) SetupWithManager(mgr ctrl.Manager) {
 	mgr.GetWebhookServer().Register("/mutate-apps-v1-statefulset", &admission.Webhook{Handler: m})
 }
@@ -97,7 +91,7 @@ func (m *StatefulsetMutator) Handle(ctx context.Context, req admission.Request) 
 		}
 
 		if groupConfig.ReasonRequired {
-			reason, hasReason := sfs.Annotations["reason"]
+			reason, hasReason := sfs.Annotations[accessReasonAnnotation]
 			if !hasReason || reason == "" {
 				return admission.Denied(fmt.Sprintf("reason is required for access group: %q", group))
 			}
@@ -174,17 +168,17 @@ func (m *StatefulsetMutator) Handle(ctx context.Context, req admission.Request) 
 	// TODO: Use Dapla Team API for this?
 	if sfs.Annotations[mountStandardBucketsAnnotation] == "true" {
 		team := group
-		var sourceData bool
+		var gc *AccessGroupConfig
 		for _, config := range m.GroupConfigs {
-			if strings.HasSuffix(group, config.Name) {
-				team = strings.TrimSuffix(group, config.Name)
-				sourceData = config.ReasonRequired
+			if team = strings.TrimSuffix(group, config.Name); team != group {
+				team = strings.TrimSuffix(team, "-")
+				gc = &config
 				break
 			}
 		}
 
 		if team != group {
-			if err := m.addStandardBuckets(ctx, team, sourceData, bucketMounts); err != nil {
+			if err := m.addStandardBuckets(ctx, team, *gc, bucketMounts); err != nil {
 				log.Error(err, "failed to add standard buckets")
 			}
 		} else {
@@ -281,7 +275,7 @@ func getFolderWithDisplayName(it *rm.FolderIterator, displayName string) (*rmpb.
 	}
 }
 
-func (m *StatefulsetMutator) addStandardBuckets(ctx context.Context, team string, sourceData bool, bucketMounts map[string]string) error {
+func (m *StatefulsetMutator) addStandardBuckets(ctx context.Context, team string, gc AccessGroupConfig, bucketMounts map[string]string) error {
 	teamFolderIt := m.Folders.ListFolders(ctx, &rmpb.ListFoldersRequest{
 		Parent: fmt.Sprintf("folders/%s", m.TeamsFolderNumber),
 	})
@@ -293,13 +287,17 @@ func (m *StatefulsetMutator) addStandardBuckets(ctx context.Context, team string
 		return fmt.Errorf("get folder %q: %w", team, err)
 	}
 
-	baseProjectName := team
-	if sourceData {
-		baseProjectName = fmt.Sprintf("%s-kilde", baseProjectName)
+	tpl, err := template.New("").Parse(gc.ProjectTemplate)
+	if err != nil {
+		return fmt.Errorf("parse template %q: %w", gc.ProjectTemplate, err)
+	}
+	projectName := strings.Builder{}
+	if err = tpl.Execute(&projectName, ProjectTemplateData{TeamName: team, Stage: m.Stage}); err != nil {
+		return fmt.Errorf("execute template %s: %w", gc.ProjectTemplate, err)
 	}
 
 	projectIt := m.Projects.SearchProjects(ctx, &rmpb.SearchProjectsRequest{
-		Query: fmt.Sprintf(`parent=%s AND state=ACTIVE AND displayName=%s`, folder.Name, fmt.Sprintf("%s-%s", baseProjectName, string(m.Stage[0]))),
+		Query: fmt.Sprintf(`parent=%s AND state=ACTIVE AND displayName=%s`, folder.Name, projectName.String()),
 	})
 
 	project, err := projectIt.Next()